separated out medication base classes into utils; cleans up med-variable file; removed borderline unused helpers
